fix: update typehinting and add method existence checks
Updated typehinting on `Output` property in `CLI` class to make it nullable and explicitly be of
`OutputInterface` type. Added method existence checks before calling methods `addEvent`,
`getVersions`, and `getLines` in `Composer` and `Web` classes to prevent potential errors and
unexpected behaviour. This results in safer code and improves resilience against possible changes
or inconsistencies in the objects these methods are supposed to exist on.

Related: quiqqer/composer#14